Refactor so that we have all the map details in one class
