+++ MuseMapalyzr/config/unranked_config.yaml
@@ -7,6 +7,8 @@
 
 # Note that some of these aren't used and the ranked config is used (e.g., sample window threshold etc.)
 
+# V8.1
+
 # Normal Sized Map Threshold
 NormalSizedMapThreshold: 0
