+++ MuseMapalyzr/strategies/varying-stacks.cs
@@ -1,3 +1,6 @@
+using System;
+using System.Collections.Generic;
+
 namespace MuseMapalyzr
 {
     public class VaryingStacksCheckSegment : CheckSegmentStrategy
@@ -132,22 +135,13 @@ namespace MuseMapalyzr
             double weightedAverage;
             if (multipliers.Count > 5 || Pattern.TotalNotes > 20) // Hard coded for now. Numbers are kinda arbitrary.
             {
-                weightedAverage = DifficultyCalculation.WeightedAverageOfValues(multipliers, 0.3, 0.9, 0.1);
+                weightedAverage = Utils.WeightedAverageOfValues(multipliers, 0.3, 0.9, 0.1);
             }
             else
             {
-                weightedAverage = DifficultyCalculation.WeightedAverageOfValues(multipliers, 0.3, 0.6, 0.4);
+                weightedAverage = Utils.WeightedAverageOfValues(multipliers, 0.3, 0.6, 0.4);
             }
             return weightedAverage;
         }
-
-
-        // public override double CalcPatternMultiplier(bool ranked)
-        // {
-        //     double nps = Pattern.Segments[0].NotesPerSecond;
-        //     double multiplier = PatternMultiplier.VaryingStacksMultiplier(nps, ranked); // assuming you have a method for this
-        //     // Console.WriteLine($"VaryingStacksCalcPatternMultiplier: {multiplier}");
-        //     return multiplier;
-        // }
     }
 }
